Module Awa_mirage.Make

SSH module given a flow

Parameters

F : Mirage_flow.S

M : Mirage_clock.MCLOCK

Signature

module FLOW = F

type error = [

| `Msg of string

| `Read of F.error

| `Write of F.write_error

]

possible errors: incoming alert, processing failure, or a problem in the underlying flow.

type write_error = [

| `Closed

| error

]

The type for write errors.

we provide the FLOW interface

include Mirage_flow.S with type Make.error := error and type Make.write_error := write_error

val pp_error : error Fmt.t

val pp_write_error : write_error Fmt.t

type flow

val read : flow -> (Cstruct.t Mirage_flow.or_eof, error) Stdlib.result Lwt.t

val write : flow -> Cstruct.t -> (unit, write_error) Stdlib.result Lwt.t

val writev : flow -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

val close : flow -> unit Lwt.t

val client_of_flow : ?⁠authenticator:Awa.Keys.authenticator -> user:string -> Awa.Hostkey.priv -> Awa.Ssh.channel_request -> FLOW.flow -> (flow, error) Stdlib.result Lwt.t

client_of_flow ~authenticator ~user key channel_request flow upgrades the existing connection to SSH, mutually authenticates, opens a channel and sends the channel request.